Homework #4
This homework focused on building a user database and then accessing it through the Python-MySQL connector. I gave the user 2 options: 1) to create a new account in the database 2) to login in using his credentials. The passwords in the database were stored as hashes by using hashlib, a Python plugin.